+++ services/data-record-service.js
@@ -9,7 +9,6 @@ const {BATCH} = require("../crdc-datahub-database-drivers/constants/batch-consta
 const {getCurrentTime} = require("../crdc-datahub-database-drivers/utility/time-utility");
 const {getFormatDateStr} = require("../utility/string-util.js")
 const {arrayOfObjectsToTSV} = require("../utility/io-util.js")
-const DataModelService = require('./dataModelService');
 const BATCH_SIZE = 300;
 const ERROR = "Error";
 const WARNING = "Warning";
@@ -57,7 +56,7 @@ const DATA_SHEET = {
     MD5SUM: "md5sum"
 };
 class DataRecordService {
-    constructor(dataRecordsCollection, dataRecordArchiveCollection, releaseCollection, fileQueueName, metadataQueueName, awsService, s3Service, qcResultsService, exportQueue, dataModelService) {
+    constructor(dataRecordsCollection, dataRecordArchiveCollection, releaseCollection, fileQueueName, metadataQueueName, awsService, s3Service, qcResultsService, exportQueue) {
         this.dataRecordsCollection = dataRecordsCollection;
         this.fileQueueName = fileQueueName;
         this.metadataQueueName = metadataQueueName;
@@ -67,7 +66,6 @@ class DataRecordService {
         this.qcResultsService = qcResultsService;
         this.exportQueue = exportQueue;
         this.releaseCollection = releaseCollection;
-        this.dataModelService = dataModelService;
     }
 
     async submissionStats(aSubmission) {
@@ -889,67 +887,6 @@ class DataRecordService {
             }
         }]);
         return genomicInfos.length > 0 ?  genomicInfos : [];
-    }
-
-    async getPropsForSubmissionAndType(submission, type) {
-        const properties = []
-        
-        const {
-            dataCommons,
-            modelVersion
-        } = submission;
-        // 1) get defined properties
-        const modelProps = await this.dataModelService.
-            getDefinedPropsByDataCommonAndType(dataCommons, modelVersion, type);
-        if (!modelProps || modelProps.length === 0) {
-            return null;
-        }
-        const modelPropNames = modelProps.map(prop => prop.handle);
-        // 2) find properties names from the first record of the submissionID and nodeType
-        const dataRecords = await this.dataRecordsCollection.aggregate([{
-            $match: {
-                submissionID: submission._id,
-                nodeType: type
-            }
-        }, {$limit: 1}]);
-        let nodeProps = [];
-        if (dataRecords.length > 0) {
-            nodeProps = Object.keys(dataRecords[0].props);
-        }
-        // 3) find node properties that are defined in the model
-        const dataModelDefinedGroup = nodeProps.filter(prop => modelPropNames.includes(prop)).map(prop => {
-            return {
-                "name": prop,
-                "group": "Data Model Defined"
-            };
-        });
-        properties.push(...dataModelDefinedGroup);
-        // 4) find node properties that are not defined in the model
-        const dataModelNotDefinedGroup = nodeProps.filter(prop => !modelPropNames.includes(prop));
-        const otherPropsGroup = dataModelNotDefinedGroup.filter(prop => prop.toLowerCase() !== "crdc_id").map(prop => {
-            return {
-                "name": prop,
-                "group": "Others"
-            };
-        });
-        properties.push(...otherPropsGroup);
-        // 5) get generated properties
-        if(dataModelNotDefinedGroup.find(prop => prop.toLowerCase() === "crdc_id")){
-            properties.push({
-                "name": "crdc_id",
-                "group": "Internal"
-            });
-        }
-        const generatedPropsArray = dataRecords[0]?.generatedProps;
-        const generatedPropNames = Object.keys(generatedPropsArray || {});
-        if (generatedPropNames.length > 0) {
-            const generatedProps = generatedPropNames.map(p => ({
-                "name": p,
-                "group": "Internal"
-            }));
-            properties.push(...generatedProps);
-        }
-        return properties.length > 0 ? properties : null;
     }
 }
